requirements.txt
python-dotenv==1.0.0
openai==1.6.1
PyPDF2==3.0.1
Pillow==10.1.0
flask==3.0.0
